// src/models/response/TaskItemResponse.ts
export interface TaskItemResponse {
    UID: string;
    TaskName: string;
    Priority: string;
    Initiator: string;
    OrganizationExecutor: string;
    Executor: string;
    DateOfCreation: string;
    DateOfCompletion: string;
    CurrentStage: string;
    Service: string;
    OrganizationClient: string;
}